[pirc/new] fix function decl.
git-svn-id: https://svn.parrot.org/parrot/trunk@31493 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ compilers/pirc/new/pirsymbol.h
@@ -93,7 +93,7 @@ unsigned find_local_label(struct lexer_state * const lexer, char * const label);
 
 unsigned get_hashcode(char * const str, unsigned num_buckets);
 
-bucket *get_bucket(hashtable * const table, unsigned hash);
+bucket *get_bucket(hashtable * const table, unsigned long hash);
 
 
 #endif /* PARROT_PIR_PIRSYMBOL_H_GUARD */
